Fixing deserialization issues with read-access when json renamed to same property name
